+++ CHANGELOG.md
@@ -3,3 +3,5 @@ All notable changes to this project will be documented in this file.
 
 #### Current Version
 
+* add support for two coin types: `0` and `145`
+* `NetworkType` enum now has `coinType` associated value in `main` case (**non-back-compatible api change**)